requirements.txt
Django>=4.0.1,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
django-cors-headers>=3.13.0, <3.14
boto3
django-dotenv
django-filter
django-storages
djangorestframework-simplejwt
gunicorn
whitenoise
dj-database-url